将`adjust`修改为`set`。

+++ components/store-rest/src/main/java/org/mallfoundry/rest/catalog/product/ProductResponse.java
@@ -90,13 +90,13 @@ public class ProductResponse implements ImmutableProductResponse {
     }
 
     @Override
-    public long getTotalSales() {
-        return 0;
+    public Long getTotalSales() {
+        return (long) 0;
     }
 
     @Override
-    public long getMonthlySales() {
-        return 0;
+    public Long getMonthlySales() {
+        return (long) 0;
     }
 
     @Override
